BUILDTOOLS-425: Disable commit.gpgSign in applyPatches.sh
If commit.gpgSign is true globally then that also applies to this
script, and so it will try to get permission to sign every commit it
applies. This means that users using BuildTools will have to temporarily
unset gpgSign from their global git config every time they want to run
BuildTools.

By setting it to false locally this is prevented.